% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plots.R
\name{add_cumsum}
\alias{add_cumsum}
\title{add_cumsum
Add columns with the cumulative sum of BP position for plotting}
\usage{
add_cumsum(coloc_summary, position = "position", chr = "CHR_var")
}
\arguments{
\item{coloc_summary}{Output from genepicoloc pipeline (summary.RDS)}
}
\description{
add_cumsum
Add columns with the cumulative sum of BP position for plotting
}
