% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{ARIC_pGWAS_annotation}
\alias{ARIC_pGWAS_annotation}
\title{ARIC pGWAS annotation}
\usage{
ARIC_pGWAS_annotation(annotation_file, coloc_out)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
ARIC pGWAS annotation
}
\examples{
Under development
}
