% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_bin}
\alias{check_bin}
\title{Check if a binary is available in the system path}
\usage{
check_bin(bin_name)
}
\arguments{
\item{bin_name}{Name or path of the binary to check}
}
\value{
TRUE if the binary is available, otherwise stops with an error
}
\description{
Check if a binary is available in the system path
}
\keyword{internal}
