% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sumstats.R
\name{check_sumstats}
\alias{check_sumstats}
\title{check_sumstats
check input sumstats for possible issues}
\usage{
check_sumstats(sumstats, Name = "Name")
}
\description{
check_sumstats
check input sumstats for possible issues
}
