% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sumstats.R
\name{check_sumstats_full}
\alias{check_sumstats_full}
\title{check_sumstats_full
check input sumstats for possible issues}
\usage{
check_sumstats_full(
  sumstats,
  fix = T,
  AF = "AF",
  BETA = "BETA",
  SE = "SE",
  P = "P",
  Name = "Name"
)
}
\description{
check_sumstats_full
check input sumstats for possible issues
}
