% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{collect_job_outputs}
\alias{collect_job_outputs}
\title{Collect and consolidate output files from job subdirectories}
\usage{
collect_job_outputs(dir_out, n_jobs, verbose = TRUE)
}
\arguments{
\item{dir_out}{Main output directory containing job subdirectories}

\item{n_jobs}{Number of job subdirectories to process}

\item{verbose}{Logical. Print progress messages}
}
\description{
Collect and consolidate output files from job subdirectories
}
