% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_coloc.R
\name{coloc_out_summary}
\alias{coloc_out_summary}
\title{coloc_out_summary}
\usage{
coloc_out_summary(
  coloc_out_all,
  output_folder,
  remove_dirname = T,
  PP.H4.abf_filt = 0.5,
  PP.H3.abf_filt = NULL
)
}
\description{
coloc_out_summary
}
