% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_over_args.R
\name{coloc_out_wrapper}
\alias{coloc_out_wrapper}
\title{coloc_out_wrapper}
\usage{
coloc_out_wrapper(
  args_list,
  output_folder = "output",
  mc_cores = 10,
  dry_run = F,
  debug_mode = F
)
}
\description{
coloc_out_wrapper
}
