% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{coloc_wrapper}
\alias{coloc_wrapper}
\title{Query sumstats and run coloc, supports multithreading / slurm.}
\usage{
coloc_wrapper(args_list, ..., do_process_wrapper = T)
}
\arguments{
\item{args_list}{list of input arguments for coloc
usually created using 'create_coloc_params_df' function}

\item{ellipsis}{with additional arguments (e.g., annotation files).}
}
\value{
results of coloc.abf function
}
\description{
Query sumstats and run coloc, supports multithreading / slurm.
}
\examples{
run_coloc()
}
