% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R, R/create_args_df.R
\name{create_args_df}
\alias{create_args_df}
\title{create_args_df
create a data.frame with arguments to be passed to coloc_wrapper}
\usage{
create_args_df(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_function,
  sumstats_1_type,
  sumstats_1_sdY,
  sumstats_2_files,
  sumstats_2_function,
  sumstats_2_type,
  sumstats_2_sdY
)

create_args_df(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_function,
  sumstats_1_type,
  sumstats_1_sdY,
  sumstats_2_files,
  sumstats_2_function,
  sumstats_2_type,
  sumstats_2_sdY
)
}
\description{
create_args_df
create a data.frame with arguments to be passed to coloc_wrapper

create_args_df
create a data.frame with arguments to be passed to coloc_wrapper
}
