% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepicoloc_susie.R
\name{create_LD_matrix}
\alias{create_LD_matrix}
\title{LD_matrix}
\usage{
create_LD_matrix(
  sumstats,
  bfile,
  fixref = F,
  keep = NULL,
  plink2 = "plink2",
  Name_var = "Name",
  A1_var = "A1"
)
}
\description{
LD_matrix
}
