% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{create_result_template}
\alias{create_result_template}
\title{Create a template for colocalization results}
\usage{
create_result_template(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_max_nlog10P,
  sumstats_2_file,
  sumstats_2_max_nlog10P,
  nsnps = NA,
  tabix = NA,
  sumstats_1_QC = NA,
  sumstats_2_QC = NA,
  coloc = NA
)
}
\arguments{
\item{CHR_var}{Chromosome}

\item{BP_START_var}{Start position}

\item{BP_STOP_var}{End position}

\item{sumstats_1_file}{Primary summary statistics file}

\item{sumstats_1_max_nlog10P}{Maximum -log10(p-value) in primary data}

\item{sumstats_2_file}{Secondary summary statistics file}

\item{sumstats_2_max_nlog10P}{Maximum -log10(p-value) in secondary data}

\item{nsnps}{Number of SNPs in analysis}

\item{tabix}{Tabix status}

\item{sumstats_1_QC}{QC status of primary data}

\item{sumstats_2_QC}{QC status of secondary data}

\item{coloc}{Colocalization status}
}
\value{
Data frame template for colocalization results
}
\description{
Create a template for colocalization results
}
