% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{create_sumstats_1_args}
\alias{create_sumstats_1_args}
\title{create_sumstats_1_args}
\usage{
create_sumstats_1_args(
  sumstats_1_file,
  sumstats_1_function = "query_sumstats_1",
  sumstats_1_type,
  sumstats_1_sdY
)
}
\description{
create_sumstats_1_args
}
