% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Package startup message}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Library path where package is installed}

\item{pkgname}{Package name}
}
\description{
Function called when the package is attached. Displays package
version and citation information.
}
\keyword{internal}
