% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onLoad}
\alias{.onLoad}
\title{Package initialization}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Library path where package is installed}

\item{pkgname}{Package name}
}
\description{
Function called when the package is loaded. Registers S3 methods
to ensure proper method dispatch.
}
\keyword{internal}
