% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{FinnGen_r9_annotation}
\alias{FinnGen_r9_annotation}
\title{FinnGen r9 annotation}
\usage{
FinnGen_r9_annotation(annotation_file, coloc_out, ...)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
FinnGen r9 annotation
}
\examples{
Under development
}
