% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flip_alleles}
\alias{flip_alleles}
\title{Flip alleles according to complementary base pairing}
\usage{
flip_alleles(vec)
}
\arguments{
\item{vec}{Character vector of nucleotide alleles (A, T, C, or G)}
}
\value{
Character vector with complementary alleles
}
\description{
Converts nucleotide alleles to their complementary bases
(A<->T, C<->G). Useful for strand flipping in genetic data.
}
\examples{
\dontrun{
flip_alleles(c("A", "T", "C", "G")) # Returns c("T", "A", "G", "C")
}
}
\keyword{internal}
