% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{format_CKDGen_r4}
\alias{format_CKDGen_r4}
\title{Format CKDGen r4 summary statistics}
\usage{
format_CKDGen_r4(sumstats)
}
\arguments{
\item{sumstats}{Data frame with CKDGen r4 summary statistics}
}
\value{
Data frame with standardized column names
}
\description{
Formats CKDGen r4 summary statistics to standard column names
}
\keyword{internal}
