% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_wrapper.R
\name{format_coloc_output}
\alias{format_coloc_output}
\title{Format the output of coloc_wrapper function.}
\usage{
format_coloc_output(coloc_output, coloc_template, N_top_SNPs = 5)
}
\arguments{
\item{coloc_output}{output of coloc::coloc.abf function.}

\item{coloc_template}{template created by genepicoloc::out_template function.}

\item{N_top_SNPs}{Number of SNPs with highest PP.H4 to output.}
}
\value{
formatted results of coloc.abf function
}
\description{
Format the output of coloc_wrapper function.
}
