% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{format_default_output}
\alias{format_default_output}
\title{Helper function to format default output (backward compatibility)}
\usage{
format_default_output(nearest_genes, pos)
}
\arguments{
\item{nearest_genes}{Data frame with nearest genes}

\item{pos}{Numeric. Position}
}
\value{
Character string in default format
}
\description{
Helper function to format default output (backward compatibility)
}
