% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{format_eQTLGen}
\alias{format_eQTLGen}
\title{Format eQTLGen summary statistics}
\usage{
format_eQTLGen(sumstats)
}
\arguments{
\item{sumstats}{Data frame with eQTLGen summary statistics}
}
\value{
Data frame with standardized column names
}
\description{
Formats eQTLGen summary statistics to standard column names
}
\keyword{internal}
