% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{format_FinnGen_r9}
\alias{format_FinnGen_r9}
\title{Format FinnGen r9 summary statistics}
\usage{
format_FinnGen_r9(sumstats)
}
\arguments{
\item{sumstats}{Data frame with FinnGen r9 summary statistics}
}
\value{
Data frame with standardized column names
}
\description{
Formats FinnGen r9 summary statistics to standard column names
}
\keyword{internal}
