% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_wrapper.R
\name{format_for_coloc}
\alias{format_for_coloc}
\title{Format sumstats data.frame for the coloc.abf input}
\usage{
format_for_coloc(sumstats_df, sumstats_type, sumstats_sdY)
}
\arguments{
\item{sumstats_df}{data frame.}

\item{sumstats_type}{quant or cc.}

\item{sumstats_sdY}{standard deviation for quantitative traits}
}
\description{
Format sumstats data.frame for the coloc.abf input
}
