% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{format_GCKD_mGWAS}
\alias{format_GCKD_mGWAS}
\title{Format GCKD mGWAS summary statistics}
\usage{
format_GCKD_mGWAS(sumstats)
}
\arguments{
\item{sumstats}{Data frame with GCKD mGWAS summary statistics}
}
\value{
Data frame with standardized column names
}
\description{
Formats GCKD mGWAS summary statistics to standard column names
}
\keyword{internal}
