% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{format_GTEXv8}
\alias{format_GTEXv8}
\title{Format GTEx v8 summary statistics}
\usage{
format_GTEXv8(sumstats)
}
\arguments{
\item{sumstats}{Data frame with GTEx v8 summary statistics}
}
\value{
Data frame with standardized column names
}
\description{
Formats GTEx v8 summary statistics to standard column names
}
\keyword{internal}
