% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{format_Kidney_eQTL}
\alias{format_Kidney_eQTL}
\title{Format Kidney eQTL summary statistics}
\usage{
format_Kidney_eQTL(sumstats)
}
\arguments{
\item{sumstats}{Data frame with Kidney eQTL summary statistics}
}
\value{
Data frame with standardized column names
}
\description{
Formats Kidney eQTL summary statistics to standard column names
}
\keyword{internal}
