% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{format_UKB_PPP_EUR}
\alias{format_UKB_PPP_EUR}
\title{Format UK Biobank PPP EUR summary statistics}
\usage{
format_UKB_PPP_EUR(sumstats)
}
\arguments{
\item{sumstats}{Data frame with UKB PPP EUR summary statistics}
}
\value{
Data frame with standardized column names
}
\description{
Formats UK Biobank PPP EUR summary statistics to standard column names
}
\keyword{internal}
