% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepicoloc_susie.R
\name{genepicoloc_susie}
\alias{genepicoloc_susie}
\title{genepicoloc_susie_rss}
\usage{
genepicoloc_susie(
  raw_gt,
  pheno,
  pheno_name,
  coverage = 0.95,
  scale_genotypes = F
)
}
\description{
genepicoloc_susie_rss
}
