% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genepicoloc_susie.R
\name{genepicoloc_susie_rss}
\alias{genepicoloc_susie_rss}
\title{genepicoloc_susie_rss}
\usage{
genepicoloc_susie_rss(
  sumstats,
  LD_matrix,
  Name_name = "Name",
  BETA_name = "BETA",
  SE_name = "SE",
  N_name = "N",
  N,
  coverage = 0.95,
  ...
)
}
\description{
genepicoloc_susie_rss
}
