% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_annotation.R
\name{get_ensembl_release}
\alias{get_ensembl_release}
\title{Get Current Ensembl Release Information}
\usage{
get_ensembl_release(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. Print release information (default: TRUE)}
}
\value{
List with release information, or NULL if unable to fetch
}
\description{
Fetch information about the current Ensembl release to help users
determine if they need to update their gene annotations.
}
\examples{
\dontrun{
# Check current Ensembl release
release_info <- get_ensembl_release()
print(release_info$release)
}

}
