% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{GTEXv8_annotation}
\alias{GTEXv8_annotation}
\title{GTEXv8 annotation}
\usage{
GTEXv8_annotation(annotation_file, coloc_out)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with processed annotation file.
}
\description{
GTEXv8 annotation
}
\examples{
Under development
}
