% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{launch_coloc_viewer}
\alias{launch_coloc_viewer}
\title{Launch GWAS Colocalization Network Viewer}
\usage{
launch_coloc_viewer(port = 3838, STUDY_BASE_PATH, available_studies)
}
\arguments{
\item{port}{Port number (default: 3838)}

\item{available_studies}{Named list of available studies or path to file containing the list}

\item{study_base_path}{Base path for study data}

\item{enable_llm}{Enable AI/LLM features (default: FALSE)}
}
\description{
Launch GWAS Colocalization Network Viewer
}
