% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_to_df_eQTL}
\alias{list_to_df_eQTL}
\title{list_to_df_eQTL
processing eQTL datasets (more than 1 gene in the same sumstats file)}
\usage{
list_to_df_eQTL(sumstats, gene, Phenotype_col = "Phenotype")
}
\description{
list_to_df_eQTL
processing eQTL datasets (more than 1 gene in the same sumstats file)
}
