% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graphics.R
\name{manhattan.plot}
\alias{manhattan.plot}
\title{manhattan.plot}
\usage{
manhattan.plot(
  sumstats,
  manhattan_title = NA,
  manhattan_subtitle = NA,
  position = "position",
  chr = "chr",
  pval = "pval",
  SNP = "SNP",
  pval_filtering,
  pval_threshold,
  nlog10 = F,
  annotate_all = F
)
}
\description{
manhattan.plot
}
