% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_over_args.R
\name{map_over_args}
\alias{map_over_args}
\title{map_over_args}
\usage{
map_over_args(
  args_df,
  mc_cores = 10,
  dry_run = F,
  debug_mode = F,
  verbose = T,
  do_rbind = T,
  save_tmp = F
)
}
\description{
map_over_args
}
