% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{match_rs}
\alias{match_rs}
\title{match_rs}
\usage{
match_rs(
  dbSNP_file,
  sumstats,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  A1_name,
  A2_name,
  SNP_name,
  ...
)
}
\description{
under development
}
