% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{mGWAS_annotation}
\alias{mGWAS_annotation}
\title{mGWAS annotation}
\usage{
mGWAS_annotation(annotation_file, coloc_out, ...)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
mGWAS annotation
}
\examples{
Under development
}
