% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{Name_by_position}
\alias{Name_by_position}
\title{Find name by position}
\usage{
Name_by_position(
  sumstats,
  tmp_name = NULL,
  CHR_name = "CHR_hg38",
  POS_name = "POS_hg38",
  A1_name = "A1_hg38",
  A2_name = "A2_hg38",
  Name_out = "Name_hg38",
  rs_name = "rs",
  unique_ID_name = "unique_ID",
  tabix_bin,
  dbSNP_file,
  do_soring = T,
  mc_cores = 4
)
}
\description{
Find name by position
}
