% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{olink_annotation}
\alias{olink_annotation}
\title{olink annotation}
\usage{
olink_annotation(olink_protein_map_3k_v1_file, coloc_out)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with processed annotation file.
}
\description{
olink annotation
}
\examples{
Under development
}
