% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{Olink_pGWAS}
\alias{Olink_pGWAS}
\title{Olink_pGWAS annotation}
\usage{
Olink_pGWAS(annotation_file, coloc_out)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
Olink_pGWAS annotation
}
\examples{
Under development
}
