% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coloc.R
\name{out_template}
\alias{out_template}
\title{Template for coloc output
under development}
\usage{
out_template(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_max_nlog10P,
  sumstats_2_file,
  sumstats_2_max_nlog10P,
  nsnps = NA
)
}
\description{
Template for coloc output
under development
}
