% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_plots.R
\name{plot_FinnGen}
\alias{plot_FinnGen}
\title{plot_FinnGen
Visualize colocalization results with FinnGen study}
\usage{
plot_FinnGen(
  coloc_summary_merged,
  FinnGen_r9_var = "FinnGen_r9",
  chr = "CHR_var",
  max_char = 30
)
}
\arguments{
\item{coloc_summary_merged}{Output from add_cumsum function}
}
\description{
plot_FinnGen
Visualize colocalization results with FinnGen study
}
