% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{prepare_coloc_dataset}
\alias{prepare_coloc_dataset}
\title{Format sumstats data.frame for the coloc.abf input}
\usage{
prepare_coloc_dataset(sumstats, sumstats_type, sumstats_sdY)
}
\arguments{
\item{sumstats}{Summary statistics object}

\item{sumstats_type}{Type of summary statistics ('quant' or 'cc')}

\item{sumstats_sdY}{Standard deviation for quantitative trait}
}
\value{
List formatted for coloc.abf input
}
\description{
Format sumstats data.frame for the coloc.abf input
}
