% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{process_sumstats_1}
\alias{process_sumstats_1}
\title{process sumstats}
\usage{
process_sumstats_1(
  sumstats_file,
  sumstats_name,
  Name_name,
  rsID_name = NULL,
  CHR_name,
  POS_name,
  A1_name,
  A2_name,
  BETA_name,
  SE_name,
  p_value_name,
  AF_name,
  N_name,
  p_threshold,
  format_columns = F
)
}
\description{
under development
}
