% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{proteomics_annotation}
\alias{proteomics_annotation}
\title{proteomics annotation}
\usage{
proteomics_annotation(
  study,
  annotation_file,
  sumstats_file = "sumstats_2_file",
  coloc_out,
  CHR_var = "CHR_var",
  BP_START_var = "BP_START_var",
  BP_STOP_var = "BP_STOP_var"
)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
proteomics annotation
}
\examples{
Under development
}
