% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_query_sumstats.R
\name{query_dbSNP}
\alias{query_dbSNP}
\title{Query dbSNP_file to get REF and ALT}
\usage{
query_dbSNP(dbSNP_file, CHR_var, BP_START_var, BP_STOP_var, rsID, ...)
}
\arguments{
\item{dbSNP_file}{path to dbSNP_file}

\item{CHR_var}{path to dbSNP_file}

\item{BP_START_var}{path to dbSNP_file}

\item{BP_STOP_var}{path to dbSNP_file}
}
\value{
data frame with rs to REF-ALT mapping
}
\description{
Query dbSNP_file to get REF and ALT
}
\examples{
under development
}
