% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_query_sumstats.R
\name{query_Ferkingstad_pGWAS}
\alias{query_Ferkingstad_pGWAS}
\title{query Ferkingstad pGWAS}
\usage{
query_Ferkingstad_pGWAS(
  sumstats_file,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  assocvariants_annotate_file
)
}
\arguments{
\item{sumstats_file}{path to sumstats sumstats.}

\item{CHR_var}{CHR (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}

\item{assocvariants_annotate_file}{FpG file with corrected maf and other QC}
}
\value{
data frame with extracted sumstats
}
\description{
Query Ferkingstad pGWAS (FpG) data to extract
every pQTL region according to GCKD
}
