% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{query_finngen_GWAS}
\alias{query_finngen_GWAS}
\title{Query finngen GWAS}
\usage{
query_finngen_GWAS(sumstats_file, CHR_var, BP_START_var, BP_STOP_var, ...)
}
\arguments{
\item{sumstats_file}{path tabix-indexed sumstats.}

\item{CHR_var}{chromosome (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}
}
\value{
data frame with extracted sumstats.
}
\description{
Query finngen GWAS data to extract a region of interest
}
\examples{
query_finngen_GWAS(sumstats_file = "finngen_R9_E4_DM2REN.gz", CHR_var = "1", BP_START_var = 100000, BP_STOP_var = 110000)
}
