% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_query_sumstats_internal.R
\name{query_GCKD_pGWAS_meta}
\alias{query_GCKD_pGWAS_meta}
\title{Query GCKD pGWAS meta}
\usage{
query_GCKD_pGWAS_meta(sumstats_file, CHR_var, BP_START_var, BP_STOP_var, ...)
}
\arguments{
\item{sumstats_file}{path tabix-indexed sumstats.}

\item{CHR_var}{chromosome (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}
}
\value{
data frame with extracted sumstats
}
\description{
Internal function, under development
}
