% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_query_sumstats.R
\name{query_sumstats_1}
\alias{query_sumstats_1}
\title{Read sumstats 1}
\usage{
query_sumstats_1(
  sumstats_file,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  ...,
  read_mode = "tabix"
)
}
\arguments{
\item{sumstats_file}{path to sumstats.}

\item{CHR_var}{chromosome (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}
}
\value{
data frame with extracted sumstats.
}
\description{
Read sumstats 1
}
