% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{query_UKB_PPP_EUR}
\alias{query_UKB_PPP_EUR}
\title{query UKB PPP pGWAS}
\usage{
query_UKB_PPP_EUR(
  sumstats_file,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  ...,
  handle_underflow = F,
  colClasses_int = c(4L, 5L),
  ncol_sumstats = 14
)
}
\arguments{
\item{sumstats_file}{path tabix-indexed sumstats.}

\item{CHR_var}{chromosome (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}
}
\value{
data frame with extracted sumstats
}
\description{
Query UKB PPP pGWAS data to extract a region of interest
}
