% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sumstats.R
\name{read_sumstats}
\alias{read_sumstats}
\title{Read and format input summary statistics (under development)}
\usage{
read_sumstats(
  sumstats,
  sumstats_file = NULL,
  Name,
  Name_new = "Name",
  A1,
  A1_new = "A1",
  BETA,
  BETA_new = "BETA",
  SE,
  SE_new = "SE",
  nlog10p_value,
  nlog10p_value_new = "nlog10P",
  rsID = NULL,
  rsID_new = "rsID",
  CHR = NULL,
  CHR_new = "CHR",
  POS = NULL,
  POS_new = "POS",
  A2 = NULL,
  A2_new = "A2",
  AF = NULL,
  AF_new = "AF",
  N = NULL,
  N_new = "N",
  p_value = NULL,
  other_columns = NULL
)
}
\arguments{
\item{sumstats}{sumstats data.frame loaded into R.}

\item{sumstats_file}{path to sumstats file.}

\item{Name}{update development}

\item{A1}{update development}

\item{BETA}{update development}

\item{SE}{update development}

\item{nlog10p_value}{update development}

\item{CHR}{update development}

\item{POS}{update development}

\item{A2}{update development}

\item{AF}{update development}

\item{N}{update development}

\item{other_columns}{vector with names of other columns}
}
\value{
data frame with formatted columns
}
\description{
Read and format input summary statistics (under development)
}
\examples{
under development
}
