% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{read_sumstats_1}
\alias{read_sumstats_1}
\title{Read sumstats 1 and format columns}
\usage{
read_sumstats_1(
  sumstats_file,
  CHR_name = "CHR",
  BP_name = "BP",
  p_value_name = "P",
  other_columns,
  read_method = "data.table"
)
}
\arguments{
\item{sumstats_file}{path to sumstats file}

\item{CHR_name}{name of the column with chromosomes}

\item{BP_name}{name of the column with positions}

\item{p_value_name}{name of the column with p-values}

\item{read_method}{either data.table (preferable, if available) or data.frame}
}
\value{
data frame with formatted columns
}
\description{
Read sumstats 1 and format columns
}
\examples{
under development
}
