% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_query_sumstats.R
\name{readtable_colCl}
\alias{readtable_colCl}
\title{readtable_colCl
Helper function to create tabix_cmd}
\usage{
readtable_colCl(ncol_sumstats, colClasses_int)
}
\description{
readtable_colCl
Helper function to create tabix_cmd
}
