% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{run_coloc}
\alias{run_coloc}
\title{Runs coloc between GCKD pQTL and every sumstats
from a list of external sumstats.}
\usage{
run_coloc(sumstats_df_1, sumstats_1_type, sumstats_df_2, sumstats_2_type)
}
\arguments{
\item{sumstats_df_1}{data frame.}

\item{sumstats_1_type}{quant by default.}

\item{sumstats_df_2}{data frame.}

\item{sumstats_2_type}{quant by default.}
}
\value{
data.table with results of coloc
}
\description{
Runs coloc between GCKD pQTL and every sumstats
from a list of external sumstats.
}
\examples{
run_coloc()
}
