% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{save_coloc_regions}
\alias{save_coloc_regions}
\title{Save coloc regions}
\usage{
save_coloc_regions(
  coloc_regions_list,
  sumstats_name,
  max_row = 1e+05,
  SKIP_name = "Name",
  CHR_place = 3,
  POS_place = 4,
  bgzip_bin = "bgzip",
  tabix_bin = "tabix",
  regions_log = "regions_log",
  coloc_regions = "coloc_regions",
  coloc_regions_PASS = "coloc_regions_PASS",
  sumstats_filt = "sumstats_filt"
)
}
\description{
Save coloc regions
}
