% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{save_job_metadata}
\alias{save_job_metadata}
\title{Save job metadata and summary information}
\usage{
save_job_metadata(
  dir_out,
  args_df,
  n_regions,
  max_regions_per_job,
  n_jobs_1,
  mc_cores,
  study_counts,
  job_times,
  verbose = TRUE
)
}
\arguments{
\item{dir_out}{Output directory path}

\item{args_df}{Arguments data frame}

\item{n_regions}{Total number of regions}

\item{max_regions_per_job}{Maximum regions per job}

\item{n_jobs_1}{Number of jobs created}

\item{mc_cores}{Number of cores used}

\item{study_counts}{Table of study counts}

\item{job_times}{Vector of job processing times}

\item{verbose}{Whether to print messages}
}
\value{
Invisibly returns list of saved file paths
}
\description{
Helper function to save comprehensive job metadata after completion
}
