% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{set_max_nlog10P}
\alias{set_max_nlog10P}
\title{Calculate and set maximum -log10(p-value) attribute}
\usage{
set_max_nlog10P(sumstats)
}
\arguments{
\item{sumstats}{A sumstats object containing a nlog10P column.}
}
\value{
The input sumstats object with an added/updated "max_nlog10P" attribute.
}
\description{
Calculates the maximum -log10(p-value) in a sumstats object and stores
it as an attribute. This is used to quickly assess whether a region
contains significant signals.
}
\details{
The maximum value is calculated excluding NA values. If the object is empty
or the nlog10P column is missing, the attribute is set to NA.
}
