% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{slurm_wrapper}
\alias{slurm_wrapper}
\title{slurm wrapper}
\usage{
slurm_wrapper(
  sumstats_1_args,
  sumstats_2_args,
  run_slurm = TRUE,
  global_objects = NULL,
  N_nodes = 20,
  N_cpus_per_node = 10,
  dry_run = T
)
}
\description{
under development
}
