% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{somascan_annotation}
\alias{somascan_annotation}
\title{somascan annotation}
\usage{
somascan_annotation(annotation_file, coloc_out)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
somascan annotation
}
\examples{
Under development
}
