% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/11_annotation.R
\name{standard_annotation}
\alias{standard_annotation}
\title{standard annotation
mGWAS, UKB TOPMed, FinnGen r9}
\usage{
standard_annotation(
  study,
  annotation_file,
  sumstats_file = "sumstats_2_file",
  coloc_out
)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
standard annotation
mGWAS, UKB TOPMed, FinnGen r9
}
\examples{
Under development
}
