% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{subset_chromosomes}
\alias{subset_chromosomes}
\title{Subset chromosomes with significant signals
to speed up computation
under development}
\usage{
subset_chromosomes(sumstats, CHR_name, nlog10p_value_name, nlogP_threshold)
}
\description{
Subset chromosomes with significant signals
to speed up computation
under development
}
