% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{subset_sumstats}
\alias{subset_sumstats}
\title{subset_sumstats}
\usage{
subset_sumstats(
  sumstats,
  coloc_regions,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  CHR_name = "CHR",
  POS_name = "POS",
  dbSNP_file = NULL,
  CHR_var_col_name = NULL,
  BP_START_var_col_name = NULL,
  BP_STOP_var_col_name = NULL,
  remove_duplicates = T,
  do_match_rs = F,
  N_mc.cores = 10,
  ...
)
}
\description{
under development
}
