% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{subset_sumstats_1}
\alias{subset_sumstats_1}
\title{subset_sumstats_1}
\usage{
subset_sumstats_1(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats,
  CHR_name = "CHR",
  POS_name = "POS",
  dbSNP_file = NULL,
  remove_duplicates = T,
  do_match_rs = F
)
}
\description{
under development
}
