% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coloc.R
\name{summarize_coloc}
\alias{summarize_coloc}
\title{Summarize coloc results}
\usage{
summarize_coloc(
  selected_studies,
  output_folder = "output",
  remove_dirname = T,
  PP.H4.abf_filt = 0.5,
  PP.H3.abf_filt = NULL,
  do_summary = T,
  do_xlsx = T
)
}
\description{
under development
}
