% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{sumstats_attr}
\alias{sumstats_attr}
\title{Add core attributes to a sumstats object}
\usage{
sumstats_attr(sumstats, sumstats_function, sumstats_type, sumstats_sdY)
}
\arguments{
\item{sumstats}{A sumstats object}

\item{sumstats_function}{Name of the function that created the data}

\item{sumstats_type}{Type of summary statistics ('quantitative' or 'binary')}

\item{sumstats_sdY}{Standard deviation for quantitative trait}
}
\value{
sumstats object with additional attributes
}
\description{
Add core attributes to a sumstats object
}
