% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{sumstats_check}
\alias{sumstats_check}
\title{Check and clean a sumstats object}
\usage{
sumstats_check(sumstats, verbose = FALSE)
}
\arguments{
\item{sumstats}{A sumstats object}

\item{verbose}{Whether to print progress messages}
}
\value{
Validated and cleaned sumstats object
}
\description{
Check and clean a sumstats object
}
