% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{sumstats_nlog10P}
\alias{sumstats_nlog10P}
\title{Calculate and set max_nlog10P attribute for a sumstats object}
\usage{
sumstats_nlog10P(sumstats)
}
\arguments{
\item{sumstats}{A sumstats object}
}
\value{
The sumstats object with updated max_nlog10P attribute
}
\description{
Calculate and set max_nlog10P attribute for a sumstats object
}
