% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{sumstats_tabix}
\alias{sumstats_tabix}
\title{Extract GWAS summary statistics from tabix-indexed file}
\usage{
sumstats_tabix(
  sumstats_file,
  coloc_regions_PASS,
  verbose = FALSE,
  test_mode = FALSE,
  file_remove = TRUE
)
}
\arguments{
\item{sumstats_file}{Path to tabix-indexed summary statistics file}

\item{coloc_regions_PASS}{Data frame with regions to extract}

\item{verbose}{Whether to print progress messages}

\item{test_mode}{If TRUE, only use the first region}

\item{file_remove}{Whether to remove temporary files}
}
\value{
A sumstats object
}
\description{
Extract GWAS summary statistics from tabix-indexed file
}
