% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{tabix_fun}
\alias{tabix_fun}
\title{tabix_fun
Helper function to create tabix_cmd}
\usage{
tabix_fun(
  sumstats_file,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  colClasses = NULL,
  sep = "\\t",
  header = T,
  show_cmd = F
)
}
\description{
tabix_fun
Helper function to create tabix_cmd
}
