% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{transcriptomics_annotation}
\alias{transcriptomics_annotation}
\title{transcriptomics annotation}
\usage{
transcriptomics_annotation(annotation_file, coloc_out, ...)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
transcriptomics annotation
}
\examples{
Under development
}
