% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_annotation.R
\name{UKB_TOPMed_annotation}
\alias{UKB_TOPMed_annotation}
\title{UKB TOPMed annotation}
\usage{
UKB_TOPMed_annotation(annotation_file, coloc_out, ...)
}
\arguments{
\item{annotation_file}{path to annotation file}
}
\value{
data.frame with annotated information.
}
\description{
UKB TOPMed annotation
}
\examples{
Under development
}
