% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_coloc.R
\name{write_coloc_out}
\alias{write_coloc_out}
\title{write_coloc_out}
\usage{
write_coloc_out(
  coloc_out,
  study,
  output_folder,
  remove_dirname = T,
  PP.H4.abf_filt = 0.5,
  PP.H3.abf_filt = NULL
)
}
\description{
write_coloc_out
}
