% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_regions.R
\name{write_regions}
\alias{write_regions}
\title{Save coloc regions}
\usage{
write_regions(
  coloc_regions_list,
  sumstats_name,
  selected_studies = NULL,
  SKIP_name = "Name",
  CHR_place = 3,
  POS_place = 4,
  bgzip_bin = "bgzip",
  tabix_bin = "tabix",
  regions_log = "regions_log",
  coloc_regions = "coloc_regions",
  coloc_regions_PASS = "coloc_regions_PASS",
  sumstats_filt = "sumstats_filt"
)
}
\description{
Save coloc regions
}
