% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebi_interface.R
\name{annotate_eQTL_Catalog}
\alias{annotate_eQTL_Catalog}
\title{annotate_eQTL_Catalog}
\usage{
annotate_eQTL_Catalog(coloc_out, annotation_df, datasets_eQTL_Catalogue)
}
\description{
annotate_eQTL_Catalog
}
