% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coloc.R
\name{coloc_wrapper}
\alias{coloc_wrapper}
\title{Query sumstats and run coloc, supports multithreading / slurm.}
\usage{
coloc_wrapper(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_function,
  ...,
  sumstats_1_type,
  sumstats_1_sdY,
  sumstats_2_file,
  sumstats_2_function,
  sumstats_2_type,
  sumstats_2_sdY,
  hyprcoloc_mode = F,
  do_process_wrapper = T,
  minP = 1e-05
)
}
\arguments{
\item{args_list}{list of input arguments for coloc
usually created using 'create_coloc_params_df' function}

\item{ellipsis}{with additional arguments (e.g., annotation files).}
}
\value{
results of coloc.abf function
}
\description{
Query sumstats and run coloc, supports multithreading / slurm.
}
\examples{
run_coloc()
}
