% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flip_alleles}
\alias{flip_alleles}
\title{flip_alleles
flip alleles according to complementary base pairing}
\usage{
flip_alleles(vec)
}
\description{
flip_alleles
flip alleles according to complementary base pairing
}
