% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{format_coloc_output}
\alias{format_coloc_output}
\title{Format the output of colocalization analysis}
\usage{
format_coloc_output(coloc_output, coloc_template, n_top_snps = 5)
}
\arguments{
\item{coloc_output}{Output from coloc::coloc.abf function}

\item{coloc_template}{Results template from create_result_template}

\item{n_top_snps}{Number of SNPs with highest PP.H4 to include}
}
\value{
Formatted results of colocalization analysis
}
\description{
Format the output of colocalization analysis
}
