% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{genepi_liftOver}
\alias{genepi_liftOver}
\title{genepi liftOver}
\usage{
genepi_liftOver(
  sumstats,
  CHR_name,
  POS_name,
  A1_name,
  A2_name,
  liftOver_bin,
  liftOver_chain_hg19ToHg38,
  dbSNP_file,
  tabix_bin,
  unique_ID_name = "unique_ID",
  mc_cores = 4,
  keep_lower = F,
  do_soring = T,
  rm_tmp_liftOver = T
)
}
\description{
genepi liftOver
}
