% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{get_coloc_regions}
\alias{get_coloc_regions}
\title{Find significant regions in sumstats and merge close regions if needed
from a list of external sumstats.}
\usage{
get_coloc_regions(
  sumstats,
  p_value_name = "P",
  CHR_name = "CHR",
  BP_name = "BP",
  p_threshold = 5e-08,
  halfwindow = 5e+05,
  log_name = "log.txt"
)
}
\arguments{
\item{sumstats}{data frame with sumstats. Mandatory columns: CHR, BP, P}

\item{p_threshold}{search for regions until no more variants below this threshold remains}

\item{log_name}{iteration log will be written to this file}
}
\value{
data frame with extracted regions
}
\description{
Find significant regions in sumstats and merge close regions if needed
from a list of external sumstats.
}
\examples{
under development
sumstats=read.table('CAD.colocalization.tsv.gz', sep='\t',header=T)
get_coloc_regions(sumstats, p_value_name = "P.value", p_threshold = 5e-100)
}
