% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_underflow}
\alias{handle_underflow}
\title{handle_underflow
Process p-values to handle underflow in case of very small p-values (p<1e-320)}
\usage{
handle_underflow(pvalue_vec, return_nlog10P = F)
}
\description{
handle_underflow
Process p-values to handle underflow in case of very small p-values (p<1e-320)
}
