% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{match_rs}
\alias{match_rs}
\title{Match rsIDs between dbSNP and sumstats}
\usage{
match_rs()
}
\arguments{
\item{dbSNP_file}{path to dbSNP_file}

\item{CHR_var}{path to dbSNP_file}

\item{BP_START_var}{path to dbSNP_file}

\item{BP_STOP_var}{path to dbSNP_file}
}
\value{
data frame with rs to REF-ALT mapping
}
\description{
Match rsIDs between dbSNP and sumstats
}
\examples{
under development
}
