% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{match_rs}
\alias{match_rs}
\title{match rs}
\usage{
match_rs(
  dbSNP_file,
  sumstats,
  CHR_var_col_name = "CHR",
  BP_START_var_col_name = "POS",
  BP_STOP_var_col_name = "POS",
  A1_name = "A1",
  A2_name = "A2",
  SNP_name = "rsID",
  window_size = 100,
  ...
)
}
\description{
match rs
}
