% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_wrapper.R, R/old.R
\name{out_template}
\alias{out_template}
\title{Create a template data.frame for coloc_wrapper output}
\usage{
out_template(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_max_nlog10P,
  sumstats_2_file,
  sumstats_2_max_nlog10P,
  nsnps = NA
)

out_template(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_max_nlog10P,
  sumstats_2_file,
  sumstats_2_max_nlog10P,
  nsnps = NA
)
}
\description{
Create a template data.frame for coloc_wrapper output

Template for coloc output
under development
}
