% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{parallel_wrapper}
\alias{parallel_wrapper}
\title{parallel wrapper}
\usage{
parallel_wrapper(
  sumstats_2_args,
  annotation_function = NULL,
  annotation_function_args = NULL,
  N_nodes = 10,
  N_cpus_per_node = 10,
  do_rbind = T,
  do_annotate = NULL,
  run_slurm = FALSE,
  global_objects = NULL,
  dry_run = T,
  debug_mode = F,
  save_RDS = T
)
}
\description{
under development
}
