% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coloc.R
\name{parallel_wrapper}
\alias{parallel_wrapper}
\title{parallel wrapper}
\usage{
parallel_wrapper(
  args_df,
  annotation_function = NULL,
  annotation_function_args = NULL,
  N_nodes = 10,
  N_cpus_per_node = 10,
  do_rbind = T,
  save_RDS_no_annotation = T,
  do_annotate = NULL,
  do_annotate_sumstats_1 = NULL,
  run_slurm = FALSE,
  global_objects = NULL,
  dry_run = T,
  debug_mode = F,
  save_RDS = T,
  minP = 1e-05
)
}
\description{
under development
}
