% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old.R
\name{parallel_wrapper}
\alias{parallel_wrapper}
\title{parallel wrapper}
\usage{
parallel_wrapper(
  args_df,
  N_cpus_per_node = 10,
  output_folder = "output",
  do_rbind = T,
  do_annotate = NULL,
  do_annotate_sumstats_1 = NULL,
  save_RDS = T,
  save_RDS_no_annotation = F,
  dry_run = T,
  debug_mode = F,
  min_nlog10P = -log10(1e-05),
  run_slurm = NULL,
  check_sumstats_2 = F
)
}
\description{
under development
}
