% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebi_interface.R
\name{query_eQTL_Catalogue}
\alias{query_eQTL_Catalogue}
\title{query_eQTL_Catalogue}
\usage{
query_eQTL_Catalogue(
  sumstats_file,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  start = 0,
  size = 1000
)
}
\description{
query_eQTL_Catalogue
}
