% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_tmp.R
\name{query_finngen_GWAS}
\alias{query_finngen_GWAS}
\title{Query UKB GWAS}
\usage{
query_finngen_GWAS(sumstats_file, CHR_var, BP_START_var, BP_STOP_var)
}
\arguments{
\item{sumstats_file}{path to UKB sumstats.}

\item{CHR_var}{CHR (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}

\item{annotation_file}{annotation file with number of samples for each GWAS.}
}
\value{
data frame with extracted sumstats.
}
\description{
Query UK Biobank GWAS data to extract a region of interest
}
