% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_query_sumstats.R
\name{query_GTEXv7_GWAS}
\alias{query_GTEXv7_GWAS}
\title{Query GTEx v7 GWAS}
\usage{
query_GTEXv7_GWAS(
  sumstats_file,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  phenotype_id_var,
  ...
)
}
\arguments{
\item{sumstats_file}{path tabix-indexed sumstats.}

\item{CHR_var}{chromosome (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}
}
\value{
data frame with extracted sumstats.
#' @examples
query_GTEXv7_GWAS(sumstats_file = "Whole_Blood.allpairs_tmp_CHR_BP_sorted.txt.gz", CHR_var = "1", BP_START_var = 1000000,
BP_STOP_var = 1010000, phenotype_id_var = "ENSG00000177757.1")
}
\description{
Query GTEx v7 GWAS data to extract a region of interest
}
