% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_query_sumstats.R
\name{query_GTEXv8_GWAS}
\alias{query_GTEXv8_GWAS}
\title{Query GTEx v8 GWAS}
\usage{
query_GTEXv8_GWAS(sumstats_file, CHR_var, BP_START_var, BP_STOP_var, ...)
}
\arguments{
\item{sumstats_file}{path tabix-indexed sumstats.}

\item{CHR_var}{chromosome (as.character "1", "2", ..., "X").}

\item{BP_START_var}{start of region, integer}

\item{BP_STOP_var}{end of region, integer}
}
\value{
data frame with extracted sumstats.
#' #' @examples
query_GTEXv8_GWAS(sumstats_file = "GTEx_V8.gz", CHR_var = "X", BP_START_var = 20000,
BP_STOP_var = 28000, phenotype_id_var = "ENSG00000167393.17")
}
\description{
Query GTEx v8 GWAS data to extract a region of interest
}
