% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_process_sumstats.R
\name{read_sumstats}
\alias{read_sumstats}
\title{Read and format input summary statistics (under development)}
\usage{
read_sumstats(
  sumstats_file,
  sumstats,
  sep_var,
  Name_name = NULL,
  Name_name_new = "Name",
  rsID_name = NULL,
  rsID_name_new = "rsID",
  CHR_name = NULL,
  CHR_name_new = "CHR",
  POS_name = NULL,
  POS_name_new = "POS",
  A1_name = NULL,
  A1_name_new = "A1",
  A2_name = NULL,
  A2_name_new = "A2",
  BETA_name = NULL,
  BETA_name_new = "BETA",
  SE_name = NULL,
  SE_name_new = "SE",
  p_value_name = NULL,
  p_value_name_new = "P",
  AF_name = NULL,
  AF_name_new = "AF",
  N_name = NULL,
  N_name_new = "N",
  other_columns = NULL
)
}
\arguments{
\item{sumstats_file}{path to sumstats file.}

\item{sumstats}{sumstats data.frame loaded into R.}

\item{sep_var}{Separator type (',', '\t', ' ', etc). Should be provided only if
read_method is read.delim (data.table usually determined separator automatically)}

\item{Name_name}{update development}

\item{CHR_name}{update development}

\item{POS_name}{update development}

\item{A1_name}{update development}

\item{A2_name}{update development}

\item{BETA_name}{update development}

\item{SE_name}{update development}

\item{p_value_name}{update development}

\item{AF_name}{update development}

\item{N_name}{update development}

\item{other_columns}{vector with names of other columns}

\item{read_method}{"read.delim" or "data.table". If data.table package is
available, "data.table" can speed up reading (usually 2-3 times faster)}
}
\value{
data frame with formatted columns
}
\description{
Read and format input summary statistics (under development)
}
\examples{
under development
}
