% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{read_sumstats_1}
\alias{read_sumstats_1}
\title{Read sumstats 1 and format columns}
\usage{
read_sumstats_1(
  sumstats_file,
  Name_name = NULL,
  rsID_name = NULL,
  CHR_name = NULL,
  POS_name = NULL,
  A1_name = NULL,
  A2_name = NULL,
  BETA_name = NULL,
  SE_name = NULL,
  p_value_name = NULL,
  AF_name = NULL,
  N_name = NULL,
  other_columns = NULL
)
}
\value{
data frame with formatted columns
}
\description{
Read sumstats 1 and format columns
}
\examples{
under development
}
