% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{run_coloc}
\alias{run_coloc}
\title{Apply "coloc" to extracted regions from 2 summary statistics}
\usage{
run_coloc(
  sumstats_1_df,
  sumstats_1_type,
  sumstats_1_sdY = NA,
  sumstats_2_df,
  sumstats_2_type,
  sumstats_2_sdY = NA
)
}
\arguments{
\item{sumstats_1_df}{data frame.}

\item{sumstats_1_type}{quant or cc.}

\item{sumstats_2_type}{quant or cc.}
}
\value{
results of coloc
}
\description{
Apply "coloc" to extracted regions from 2 summary statistics
}
\examples{
run_coloc()
}
