% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{run_coloc}
\alias{run_coloc}
\title{Run colocalization analysis using coloc.abf}
\usage{
run_coloc(sumstats_1, sumstats_2, silent = TRUE)
}
\arguments{
\item{sumstats_1}{Primary summary statistics object}

\item{sumstats_2}{Secondary summary statistics object}

\item{silent}{Whether to suppress coloc.abf output messages}
}
\value{
Results from coloc.abf function
}
\description{
Run colocalization analysis using coloc.abf
}
