% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coloc_wrapper.R
\name{run_coloc}
\alias{run_coloc}
\title{Run coloc.abf given either two data.frames:sumstats_1_df, sumstats_2_df,
or parameters to query these data.frames: CHR, BP, file, and function
sumstats_type and sumstats_sdY are mandatory}
\usage{
run_coloc(
  sumstats_1_df,
  sumstats_2_df,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats_1_file,
  sumstats_1_function,
  sumstats_2_file,
  sumstats_2_function,
  sumstats_1_type,
  sumstats_1_sdY,
  sumstats_2_type,
  sumstats_2_sdY,
  silent = T
)
}
\description{
Run coloc.abf given either two data.frames:sumstats_1_df, sumstats_2_df,
or parameters to query these data.frames: CHR, BP, file, and function
sumstats_type and sumstats_sdY are mandatory
}
