% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_coloc.R
\name{run_coloc}
\alias{run_coloc}
\title{Run coloc function using 2 extracted regions}
\usage{
run_coloc(
  sumstats_1_df,
  sumstats_1_type,
  sumstats_1_sdY,
  sumstats_2_df,
  sumstats_2_type,
  sumstats_2_sdY
)
}
\arguments{
\item{sumstats_1_df}{data frame.}

\item{sumstats_1_type}{quant or cc.}

\item{sumstats_2_type}{quant or cc.}
}
\value{
results of coloc.abf function
}
\description{
Run coloc function using 2 extracted regions
}
\examples{
run_coloc()
}
