% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{run_region}
\alias{run_region}
\title{Run colocalization for a specific genomic region}
\usage{
run_region(
  sumstats_1,
  sumstats_2,
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  min_nlog10P = 5
)
}
\arguments{
\item{sumstats_1}{Primary summary statistics object}

\item{sumstats_2}{Secondary summary statistics object}

\item{CHR_var}{Chromosome}

\item{BP_START_var}{Start position}

\item{BP_STOP_var}{End position}

\item{min_nlog10P}{Minimum -log10(p-value) threshold}
}
\value{
Data frame with colocalization results
}
\description{
Run colocalization for a specific genomic region
}
