% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coloc_regions.R
\name{subset_chromosomes}
\alias{subset_chromosomes}
\title{subset_chromosomes
Subset chromosomes with significant signals to speed up computation
Used by get_coloc_regions()}
\usage{
subset_chromosomes(sumstats, CHR_name, nlog10p_value_name, nlogP_threshold)
}
\description{
subset_chromosomes
Subset chromosomes with significant signals to speed up computation
Used by get_coloc_regions()
}
