% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{subset_sumstats_1}
\alias{subset_sumstats_1}
\title{Read sumstats 1 and format columns}
\usage{
subset_sumstats_1(sumstats, coloc_regions)
}
\arguments{
\item{sumstats}{sumstats file created by read_sumstats_1()}

\item{coloc_regions}{file created by get_coloc_regions()}
}
\value{
data frame with formatted columns
}
\description{
Read sumstats 1 and format columns
}
\examples{
under development
}
