% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_coloc.R
\name{subset_sumstats_1}
\alias{subset_sumstats_1}
\title{Read sumstats 1 and format columns}
\usage{
subset_sumstats_1(
  CHR_var,
  BP_START_var,
  BP_STOP_var,
  sumstats,
  dbSNP_file = NULL,
  remove_duplicates = T,
  do_match_rs = F
)
}
\arguments{
\item{sumstats}{sumstats file created by read_sumstats_1()}

\item{coloc_regions}{file created by get_coloc_regions()}
}
\value{
data frame with formatted columns
}
\description{
Read sumstats 1 and format columns
}
\examples{
under development
}
