% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{tabix_FinnGen_r9}
\alias{tabix_FinnGen_r9}
\title{tabix_FinnGen_r9
https://finngen.gitbook.io/documentation/data-description}
\usage{
tabix_FinnGen_r9(sumstats_file, coloc_regions_PASS)
}
\description{
tabix_FinnGen_r9
https://finngen.gitbook.io/documentation/data-description
}
