% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_sumstats_custom.R
\name{tabix_Icelanders_pGWAS}
\alias{tabix_Icelanders_pGWAS}
\title{tabix_Icelanders_pGWAS}
\usage{
tabix_Icelanders_pGWAS(sumstats_file, coloc_regions_PASS)
}
\description{
tabix_Icelanders_pGWAS
}
